+++ src/components/base/b-router/modules/transition/class.ts
@@ -20,8 +20,12 @@ import type { TransitionContext } from 'components/base/b-router/modules/transit
 const
 	$$ = symbolGenerator();
 
-const transitionLabel = {
+const transitionPerformLabel = {
 	label: $$.transition
+};
+
+const transitionExecuteLabel = {
+	label: $$.transitionExecute
 };
 
 export default class Transition {
@@ -156,13 +160,13 @@ export default class Transition {
 		this.initNewRouteInfo();
 
 		this.scroll.createSnapshot();
-		await $a.promise(this.scroll.updateCurrentRouteScroll(), transitionLabel);
+		await $a.promise(this.scroll.updateCurrentRouteScroll(), transitionExecuteLabel);
 
 		// We didn't find any route matching the given ref
 		if (this.newRouteInfo == null) {
 			// The transition was user-generated, then we need to save the scroll
 			if (!SSR && this.method !== 'event' && this.ref != null) {
-				await $a.promise(engine[this.method](this.ref, this.scroll.getSnapshot()), transitionLabel);
+				await $a.promise(engine[this.method](this.ref, this.scroll.getSnapshot()), transitionExecuteLabel);
 			}
 
 			return;
@@ -268,7 +272,7 @@ export default class Transition {
 				return;
 			}
 
-			await $a.promise(engine[this.method](newRoute.url, plainInfo), transitionLabel).then(() => {
+			await $a.promise(engine[this.method](newRoute.url, plainInfo), transitionPerformLabel).then(() => {
 				const isSoftTransition = r.route != null && Object.fastCompare(
 					router.convertRouteToPlainObjectWithoutProto(currentRoute),
 					router.convertRouteToPlainObjectWithoutProto(newRoute)